IncludeRelatedEntitiesCollection and Resource were improved and were put into ImprovedTraits directory
