+++ publisher/laravel-app/app/Http/Admin/SpecProperties/Piercings/Piercings/Controllers/PiercingJewelleryRelatedController.php
@@ -1,14 +1,14 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Admin\Piercing\Piercings\Controllers;
+namespace App\Http\Admin\SpecProperties\Piercings\Piercings\Controllers;
 
+use App\Http\Admin\Jewellery\Jewelleries\Resources\JewelleryResource;
 use App\Http\Controllers\Controller;
-use App\Http\Shared\Resources\Identifiers\ApiEntityIdentifierResource;
 use Domain\JewelleryProperties\Piercings\Piercings\Services\Relationships\PiercingJewelleryRelationshipService;
 use Illuminate\Http\JsonResponse;
 
-final class PiercingJewelleryRelationshipController extends Controller
+final class PiercingJewelleryRelatedController extends Controller
 {
     public function __construct(public PiercingJewelleryRelationshipService $service)
     {
@@ -17,6 +17,7 @@ final class PiercingJewelleryRelationshipController extends Controller
     public function index(int $id): JsonResponse
     {
         $model = $this->service->index($id);
-        return (new ApiEntityIdentifierResource($model))->response();
+
+        return (new JewelleryResource($model))->response();
     }
 }
