+++ publisher/laravel-app/app/Http/Admin/SpecProperties/Piercings/Piercings/Resources/PiercingCollection.php
@@ -1,10 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Admin\Piercing\Piercings\Resources;
+namespace App\Http\Admin\SpecProperties\Piercings\Piercings\Resources;
 
 use App\Http\Shared\Resources\Traits\IncludeRelatedEntitiesCollectionTrait;
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 
 final class PiercingCollection extends ResourceCollection
